% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_table.R
\name{add_table}
\alias{add_table}
\title{Function that adds a data frame to an (existing) .xlsx workbook sheet}
\usage{
add_table(
  Table,
  WbTitle,
  SheetTitle,
  TableTitle,
  StartRow = 1,
  StartCol = 1,
  FormatList,
  HeightTableTitle = 2,
  Footnote1,
  Footnote2,
  Footnote3
)
}
\arguments{
\item{Table}{: data frame to be exported to the workbook sheet}

\item{WbTitle}{: workbook}

\item{SheetTitle}{: string used for the sheet's name}

\item{TableTitle}{: string used for the data frame's title}

\item{StartRow}{: export start line number in the sheet (by default 1)}

\item{StartCol}{: export start column number in the sheet (by default 1)}

\item{FormatList}{: list that indicates the format of each column of the data frame}

\item{HeightTableTitle}{: multiplier (if needed) for the height of the title line (by default 2)}

\item{Footnote1}{: string for footnote1}

\item{Footnote2}{: string for footnote2}

\item{Footnote3}{: string for footnote1}
}
\value{
excel wb object
}
\description{
Function that adds a data frame to an (existing) .xlsx workbook sheet
}
