% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calcskippedrow}
\alias{calcskippedrow}
\title{Utility function that returns the number of rows that need to be skipped
to correctly account for previous data frames already present in the sheet}
\usage{
calcskippedrow(mylist, x)
}
\arguments{
\item{mylist}{a list of data frames}

\item{x}{a number}
}
\description{
Utility function that returns the number of rows that need to be skipped
to correctly account for previous data frames already present in the sheet
}
\examples{
mydf <- list(iris,cars,mtcars)
calcskippedrow(maliste=mydf,x=1) # return 0
calcskippedrow(maliste=mydf,x=2) # return 150
calcskippedrow(maliste=mydf,x=3) # return 200 (150 rows of iris + 50 of cars)
calcskippedrow(maliste=mydf,x=4) # return 232

}
\keyword{internal}
