% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calcstartrow}
\alias{calcstartrow}
\title{Utility function that returns start row to correctly separate df written in a same sheet}
\usage{
calcstartrow(x)
}
\arguments{
\item{x}{a number}
}
\description{
Utility function that returns start row to correctly separate df written in a same sheet
}
\examples{
calcstartrow(1) # return 1
calcstartrow(2) # return 11
calcstartrow(3) # return 21

}
\keyword{internal}
