% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_range_string}
\alias{convert_range_string}
\title{Utility function that takes converts a string representing
a range of numbers (e.g. "1:3") into a vector of numbers}
\usage{
convert_range_string(range_string)
}
\arguments{
\item{range_string}{A string representing a range of numbers}
}
\value{
A vector of numbers
}
\description{
Utility function that takes converts a string representing
a range of numbers (e.g. "1:3") into a vector of numbers
}
\examples{
convert_range_string("1:3")
#' # Output:
# [1] 1 2 3

}
\keyword{internal}
