% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_indices_from_vector}
\alias{get_indices_from_vector}
\title{Utility function that takes a vector as input and returns
the indices of the first and last element.}
\usage{
get_indices_from_vector(vector)
}
\arguments{
\item{vector}{a vector}
}
\description{
Utility function that takes a vector as input and returns
the indices of the first and last element.
}
\examples{
myvector <- c("momo","momo","momo","mumu","mumu")
get_indices_from_vector(myvector)
#' # Output:
# [1] "1:5"

}
\keyword{internal}
