% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_indices_of_identical_elements}
\alias{get_indices_of_identical_elements}
\title{Utility function that takes a vector as input and returns
the indices of the elements in the vector that are identical.}
\usage{
get_indices_of_identical_elements(vector)
}
\arguments{
\item{vector}{a vector}
}
\description{
Utility function that takes a vector as input and returns
the indices of the elements in the vector that are identical.
}
\examples{
myvector <- c("momo","momo","momo","mumu","mumu")
get_indices_of_identical_elements(myvector)
#' # Output:
# [1] "1:3" "4:5"

}
\keyword{internal}
