% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_toc.R
\name{insert_toc}
\alias{insert_toc}
\title{Insert Table of Contents to an existing workbook}
\usage{
insert_toc(wb, sheets)
}
\arguments{
\item{wb}{Workbook object to modify}

\item{sheets}{A named list in the format list("dataframe" = "sheet_name")}
}
\value{
Nothing, this function is called for its side effects
}
\description{
This function takes an existing workbook and a list or vector
  of sheet names, and creates a Table of Contents sheet with hyperlinks.
}
\seealso{
[toxlsx()]
}
\keyword{internal}
