% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-tablexlsx.R
\docType{package}
\name{tablexlsx-package}
\alias{tablexlsx}
\alias{tablexlsx-package}
\title{tablexlsx: Export Data Frames to Excel Workbook}
\description{
Collection of functions that allow to export data frames to excel workbook.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ddotta.github.io/tablexlsx/}
  \item \url{https://github.com/ddotta/tablexlsx}
  \item Report bugs at \url{https://github.com/ddotta/tablexlsx/issues}
}

}
\author{
\strong{Maintainer}: Damien Dotta \email{damien.dotta@live.fr}

}
\keyword{internal}
