% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxlsx.R
\name{toxlsx}
\alias{toxlsx}
\title{Convert R data frames to excel files}
\usage{
toxlsx(
  object,
  tosheet = list(),
  title = list(),
  columnstyle = list(default = NULL),
  footnote1 = list(),
  footnote2 = list(),
  footnote3 = list(),
  path,
  automaticopen = TRUE
)
}
\arguments{
\item{object}{data.frame or list to be converted to excel}

\item{tosheet}{list of sheet names for each element of object}

\item{title}{list of title for each element of object}

\item{columnstyle}{list of style for each element of object}

\item{footnote1}{list of footnote1 for each element of object}

\item{footnote2}{list of footnote2 for each element of object}

\item{footnote3}{list of footnote3 for each element of object}

\item{path}{path to save excel file}

\item{automaticopen}{logical indicating if excel file should open automatically}
}
\value{
an excel file
}
\description{
This is a function called "toxlsx" that takes in several parameters
including an "object" (which must be a data frame or list),
"tosheet", "title", "columnstyle", "footnote1", "footnote2", and "footnote3"
(all of which must be lists), and "path" and "automaticopen".
The function starts by performing several assertions to ensure that the
input parameters are of the correct type and format. It then creates an
empty list called "output" and names its elements using the "output_name"
variable. The function then loops through each element in "output_name"
and assigns values to several elements within the "output" list.
The function then creates an empty workbook and fills it with the data
from the "output" list. The function also includes some additional logic
for formatting columns in the workbook.
}
