% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDataByGroup.R
\name{writeDataByGroup}
\alias{writeDataByGroup}
\title{WriteData by groups}
\usage{
writeDataByGroup(
  wb,
  sheet,
  x,
  group = ".group",
  groupsep = " / ",
  groupname = FALSE,
  groupequal = " = ",
  startRow = 1,
  ...
)
}
\arguments{
\item{wb}{A Workbook object}

\item{sheet}{The worksheet to write to}

\item{x}{A data.frame}

\item{group}{Character vector indicating the name of the columns by which to group}

\item{groupsep}{Character vector that separates modalities if multiple columns in group}

\item{groupname}{Boolean indicating whether the name of the grouping variable should be written}

\item{groupequal}{Character vector that separates the name and the modalities of grouping variable (if groupname = TRUE)}

\item{startRow}{The starting row to write to}

\item{...}{Additional arguments passed to openxlsx::writeData()}
}
\value{
A Workbook in which data has been written by group
}
\description{
WriteData by groups
}
